requirements.txt
pip
youtube_dl
discord.py[voice] ~= 1.4.0
ffmpeg
colorlog
PyNaCl
asyncpraw
spotipy
TikTokAPI
pathlib